// src/app/models/product.ts
export interface Product {
  $key: string;
  name: string;
  category: string;
  price: number;
  imageUrl: string;
  screenDiagonal: string;
  screenResolution: string;
  images: Array<any>[];
  reviews: Array<any>[];
  discType: string;
  discCapacity: string;
  cpu: {
    name: string;
  };
  graphicIntegrated: string;
  promotion: number;
  chipset: string;
  socketRam: string;
  socketMb: string;
  socket: string;
  maxRam: string;
  model: string;
  additionalEquipment: Array<any>[];
  platform: string;
  genres: Array<any>[];
  drive: string;
  age: string;
  image: string;
  color: string;
  wifi: string;
  hdtv: string;
  psDisc: string;
  type: string;
  psGraphic: string;
  psCpu: string;
  psRam: string;
  dimensions: string;
  rotationSpeed: string;
  maxNoise: string;
  airFlow: string;
  coolerSize: string;
  cpuFrequency: string;
  frequencyTurbo: string;
  multiplier: string;
  unlocked: string;
  cores: string;
  modelCpu: string;
  windows: string;
  standardMb: string;
  crossfire: string;
  sli: string;
  destinyAntena: string;
  energyGain: string;
  polarization: string;
  connector: string;
  typeCommunication: string;
  interface: string;
  standard: string;
  frequency: string;
  maxSpeedTransmision: string;
  functions: Array<any>[];
  operationMode: Array<any>[];
  portsFront: Array<any>[];
  installedCoolers: Array<any>[];
  amount525: string;
  amount35: string;
  standards: Array<any>[];
  discRotation: string;
  widthFormat: string;
  sizeMemory: string;
  memoryFrequency: string;
  coreFrequency: string;
  typeConnection: string;
  producerChipset: string;
  modelChipset: string;
  typeMemory: string;
  strength: string;
  amountCoolers: string;
  diameterCooler:string;
  performance: string;
  standardVersion: string;
  ramMemory: string;
  singleModule: string;
  amountModule: string;
  latency: string;
  fullCapacity: string;
  ramFrequency: string;
  manufacturer: string;
  sampling: string;
  samplingFrequency: string;
  maxTransfer: string;
  tripod: string;
  band: string;
  noiseReduction: string;
  matrixType: string;
  screenType: string;
  backlight: string;
  hasSpeakers: string;
  dpi: string;
  amountButtons: string;
  signal: string;
  amountSpeaker: string;
  amountSatellite: string;
  hasSubwoofer: string;
  powerSatellite: string;
  powerCentral: string;
  language: string;
  version: string;
  timeLicense: string;
  user: string;
}